utilize the new delete code

+++ src/main/java/org/elasticsearch/indices/IndicesService.java
@@ -481,7 +481,9 @@ public class IndicesService extends AbstractLifecycleComponent<IndicesService> i
             // the store metadata gets wiped anyway even without the lock this is just best effort since
             // every shards deletes its content under the shard lock it owns.
             logger.debug("{} deleting index store reason [{}]", index, reason);
-            nodeEnv.deleteIndexDirectorySafe(index, 0, indexSettings);
+            if (canDeleteIndexContents(index, indexSettings)) {
+                nodeEnv.deleteIndexDirectorySafe(index, 0, indexSettings);
+            }
         } catch (LockObtainFailedException ex) {
             logger.debug("{} failed to delete index store - at least one shards is still locked", ex, index);
         } catch (Exception ex) {
@@ -500,13 +502,17 @@ public class IndicesService extends AbstractLifecycleComponent<IndicesService> i
      * @param indexSettings the shards index settings.
      * @throws IOException if an IOException occurs
      */
-    public void deleteShardStore(String reason, ShardLock lock, Settings indexSettings) throws IOException {
+    public void deleteShardStore(String reason, ShardLock lock, Settings indexSettings, boolean ownsShard) throws IOException {
         ShardId shardId = lock.getShardId();
-        if (canDeleteShardContent(shardId, indexSettings) == false) {
-            throw new ElasticsearchIllegalStateException("Can't delete shard " + shardId);
+        if (canDeleteShardContent(shardId, indexSettings, ownsShard) == false) {
+            if (canDeleteShardContent(shardId, indexSettings, true)) {
+                throw new ElasticsearchIllegalStateException("Can't delete shard " + shardId + " ownsShard: " + ownsShard);
+            }
+            logger.trace("{} skip deleting shard shard ownsShard: {} ", shardId, ownsShard);
+        } else {
+            logger.trace("{} deleting shard reason [{}]", shardId, reason);
+            nodeEnv.deleteShardDirectoryUnderLock(lock, indexSettings);
         }
-        logger.trace("{} deleting shard reason [{}]", shardId, reason);
-        nodeEnv.deleteShardDirectoryUnderLock(lock, indexSettings);
     }
 
     /**
@@ -520,11 +526,24 @@ public class IndicesService extends AbstractLifecycleComponent<IndicesService> i
      */
     public void deleteShardStore(String reason, ShardId shardId, IndexMetaData metaData) throws IOException {
         final Settings indexSettings = buildIndexSettings(metaData);
-        if (canDeleteShardContent(shardId, indexSettings) == false) {
+        if (canDeleteShardContent(shardId, indexSettings, false) == false) {
             throw new ElasticsearchIllegalStateException("Can't delete shard " + shardId);
         }
         nodeEnv.deleteShardDirectorySafe(shardId, indexSettings);
         logger.trace("{} deleting shard reason [{}]", shardId, reason);
+    }
+
+    //NOCOMMIT document this
+    public boolean canDeleteIndexContents(Index index, Settings indexSettings) {
+        final Tuple<IndexService, Injector> indexServiceInjectorTuple = this.indices.get(index);
+        if (indexSettings.getAsBoolean(IndexMetaData.SETTING_SHADOW_REPLICAS, false) == false) {
+            if (indexServiceInjectorTuple == null && nodeEnv.hasNodeFile()) {
+                return true;
+            }
+        } else {
+            logger.trace("{} skipping index directory deletion due to shadow replicas", index);
+        }
+        return false;
     }
 
     /**
@@ -545,18 +564,22 @@ public class IndicesService extends AbstractLifecycleComponent<IndicesService> i
         // The plan was to make it harder to miss-use and ask for metadata instead of simple settings
         assert shardId.getIndex().equals(metaData.getIndex());
         final Settings indexSettings = buildIndexSettings(metaData);
-        return canDeleteShardContent(shardId, indexSettings);
+        return canDeleteShardContent(shardId, indexSettings, false);
     }
 
-    private boolean canDeleteShardContent(ShardId shardId, @IndexSettings Settings indexSettings) {
+    private boolean canDeleteShardContent(ShardId shardId, @IndexSettings Settings indexSettings, boolean ownsShard) {
         final Tuple<IndexService, Injector> indexServiceInjectorTuple = this.indices.get(shardId.getIndex());
         // TODO add some protection here to prevent shard deletion if we are on a shard FS or have ShadowReplicas enabled.
-        if (indexServiceInjectorTuple != null && nodeEnv.hasNodeFile()) {
-            final IndexService indexService = indexServiceInjectorTuple.v1();
-            return indexService.hasShard(shardId.id()) == false;
-        } else if (nodeEnv.hasNodeFile()) {
-            final Path[] shardLocations = nodeEnv.shardDataPaths(shardId, indexSettings);
-            return FileSystemUtils.exists(shardLocations);
+        if (indexSettings.getAsBoolean(IndexMetaData.SETTING_SHADOW_REPLICAS, ownsShard) == false) {
+            if (indexServiceInjectorTuple != null && nodeEnv.hasNodeFile()) {
+                final IndexService indexService = indexServiceInjectorTuple.v1();
+                return indexService.hasShard(shardId.id()) == false;
+            } else if (nodeEnv.hasNodeFile()) {
+                final Path[] shardLocations = nodeEnv.shardDataPaths(shardId, indexSettings);
+                return FileSystemUtils.exists(shardLocations);
+            }
+        } else {
+            logger.trace("{} skipping shard directory deletion due to shadow replicas", shardId);
         }
         return false;
     }
